Fix diffusion auto-residency test placement
